+++ app/src/main/java/woowacourse/movie/ui/base/BaseView.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.view
+package woowacourse.movie.ui.base
 
 interface BaseView {
     fun handleError()
