+++ app/src/main/java/woowacourse/movie/ui/complete/MovieReservationCompleteContract.kt
@@ -1,8 +1,8 @@
-package woowacourse.movie.conrtract
+package woowacourse.movie.ui.complete
 
 import woowacourse.movie.model.Ticket
 import woowacourse.movie.model.data.dto.MovieContent
-import woowacourse.movie.view.BaseView
+import woowacourse.movie.ui.base.BaseView
 
 interface MovieReservationCompleteContract {
     interface View : BaseView {
