+++ app/src/main/java/woowacourse/movie/ui/complete/MovieReservationCompletePresenter.kt
@@ -1,6 +1,5 @@
-package woowacourse.movie.presenter
+package woowacourse.movie.ui.complete
 
-import woowacourse.movie.conrtract.MovieReservationCompleteContract
 import woowacourse.movie.model.ReservationCount
 import woowacourse.movie.model.Ticket
 import woowacourse.movie.model.data.MovieContentsImpl
