+++ app/src/main/java/woowacourse/movie/ui/main/MainActivity.kt
@@ -1,11 +1,10 @@
-package woowacourse.movie.view.activity
+package woowacourse.movie.ui.main
 
 import android.os.Bundle
 import android.widget.ListView
 import woowacourse.movie.R
-import woowacourse.movie.conrtract.MainContract
-import woowacourse.movie.presenter.MainPresenter
-import woowacourse.movie.view.adapter.MovieContentListAdapter
+import woowacourse.movie.ui.adapter.MovieContentListAdapter
+import woowacourse.movie.ui.base.BaseActivity
 
 class MainActivity : BaseActivity<MainContract.Presenter>(), MainContract.View {
     private val presenter: MainContract.Presenter by lazy { initializePresenter() }
