+++ app/src/main/java/woowacourse/movie/ui/main/MainContract.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.conrtract
+package woowacourse.movie.ui.main
 
 interface MainContract {
     interface View {
